finance-tracker/requirements.txt
apturl==0.5.2
asgiref==3.9.1
attrs==21.2.0
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.2.1
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.4
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
fasteners==0.14.1
Flask==3.1.1
folder-color==0.2.0
future==0.18.2
httplib2==0.20.2
icoextract==0.1.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==3.0.2
meld==3.20.4
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pefile==2021.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pyelftools==0.27
pygit2==1.6.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pympress==1.7.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt===2.4.0-ubuntu4-zorin1
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.3
texttable==1.6.4
typing_extensions==4.14.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
watchdog==2.1.6
websocket-client==1.2.3
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
zipp==1.0.0
zorin-appearance==5.3
zorin-connect==1.2
zorin-exec-guard==1.5
